Update README.md -Makes it work again with qb-target and more realistic
Hi, I was just looking for a mugshot script and saw that qb-target no longer works with the script. So I edited the readme.md to make it work again for you and be more realistic. Now you have to set a model hash or a propname for the qb-target // CQC mugshot. For example, Gabz-MRPD has a camera in the interrogation room at the bottom, which is indicated here in the example with the model hash. You can also place an NPC there instead and specify the ped name.
Any Questions about it? Join my DC and feel free to ask. PEACE